+++ jest.config.js
@@ -69,7 +69,14 @@ module.exports = {
 
   // A set of global variables that need to be available in all test environments
 
-  // globals: {},
+  globals: {
+    'ts-jest': {
+      tsconfig: {
+        allowJs: true,
+        sourceMap: true,
+      },
+    },
+  },
 
   // The maximum amount of workers used to run your tests. Can be specified as % or a number. E.g. maxWorkers: 10% will use 10% of your CPU amount + 1 as the maximum worker number. maxWorkers: 2 will use a maximum of 2 workers.
 
@@ -229,16 +236,13 @@ module.exports = {
   // A map from regular expressions to paths to transformers
 
   transform: {
-    '^.+\\.(ts|tsx)$': 'ts-jest',
+    '^.+\\.(ts|tsx|js)$': 'ts-jest',
   },
 
   // An array of regexp pattern strings that are matched against all source file paths, matched files will skip transformation
-
-  // transformIgnorePatterns: [
-
-  //   "\\\\node_modules\\\\"
-
-  // ],
+  transformIgnorePatterns: [
+    '<rootDir>/node_modules/(?!(react-table/src/filterTypes.js))',
+  ],
 
   // An array of regexp pattern strings that are matched against all modules before the module loader will automatically return a mock for them