+++ src/core/Table/actionHandlers/selectHandler.ts
@@ -2,8 +2,11 @@
  * Copyright (c) Bentley Systems, Incorporated. All rights reserved.
  * See LICENSE.md in the project root for license terms and full copyright notice.
  *--------------------------------------------------------------------------------------------*/
-import { TableInstance, TableState } from 'react-table';
+import { ActionType, Row, TableInstance, TableState } from 'react-table';
 
+/**
+ * Handles selection when clicked on a checkbox.
+ */
 export const onSelectHandler = <T extends Record<string, unknown>>(
   newState: TableState<T>,
   instance?: TableInstance<T>,
@@ -18,14 +21,85 @@ export const onSelectHandler = <T extends Record<string, unknown>>(
     return;
   }
 
-  const selectedData: T[] = [];
   const newSelectedRowIds = {} as Record<string, boolean>;
-  instance.initialRows.forEach((row) => {
-    if (newState.selectedRowIds[row.id] && !isRowDisabled?.(row.original)) {
-      newSelectedRowIds[row.id] = true;
-      selectedData.push(row.original);
+
+  const handleRow = (row: Row<T>) => {
+    if (isRowDisabled?.(row.original)) {
+      return false;
     }
-  });
+
+    let isAllSubSelected = true;
+    row.initialSubRows.forEach((subRow) => {
+      const result = handleRow(subRow);
+      if (!result) {
+        isAllSubSelected = false;
+      }
+    });
+
+    // If `selectSubRows` is false, then no need to select sub-rows and just check current selection state.
+    // If a row doesn't have sub-rows then check its selection state.
+    // If it has sub-rows then check whether all of them are selected.
+    if (
+      (!instance.selectSubRows && newState.selectedRowIds[row.id]) ||
+      (!row.initialSubRows.length && newState.selectedRowIds[row.id]) ||
+      (row.initialSubRows.length && isAllSubSelected)
+    ) {
+      newSelectedRowIds[row.id] = true;
+    }
+    return !!newSelectedRowIds[row.id];
+  };
+  instance.initialRows.forEach((row) => handleRow(row));
+
+  const selectedData = getSelectedData(newSelectedRowIds, instance);
+
   newState.selectedRowIds = newSelectedRowIds;
   onSelect?.(selectedData, newState);
+};
+
+/**
+ * Handles selection when clicked on a row.
+ */
+export const onSingleSelectHandler = <T extends Record<string, unknown>>(
+  state: TableState<T>,
+  action: ActionType,
+  instance?: TableInstance<T>,
+  onSelect?: (
+    selectedData: T[] | undefined,
+    tableState?: TableState<T>,
+  ) => void,
+  isRowDisabled?: (rowData: T) => boolean,
+) => {
+  const selectedRowIds = { [action.id]: true } as Record<string, boolean>;
+  if (instance?.selectSubRows) {
+    const handleRow = (row: Row<T>) => {
+      selectedRowIds[row.id] = true;
+      row.subRows.forEach((r) => handleRow(r));
+    };
+    handleRow(instance.rowsById[action.id]);
+  }
+
+  const newState = {
+    ...state,
+    selectedRowIds,
+  };
+  // Passing to `onSelectHandler` to handle filtered sub-rows
+  onSelectHandler(newState, instance, onSelect, isRowDisabled);
+
+  return newState;
+};
+
+const getSelectedData = <T extends Record<string, unknown>>(
+  selectedRowIds: Record<string, boolean>,
+  instance?: TableInstance<T>,
+) => {
+  const selectedData: T[] = [];
+  const setSelectedData = (row: Row<T>) => {
+    if (selectedRowIds[row.id]) {
+      selectedData.push(row.original);
+    }
+    row.initialSubRows.forEach((subRow) => setSelectedData(subRow));
+  };
+  instance?.initialRows.forEach((row) => setSelectedData(row));
+
+  return selectedData;
 };