+++ src/core/Table/hooks/index.ts
@@ -4,3 +4,5 @@
  *--------------------------------------------------------------------------------------------*/
 export { EXPANDER_CELL_ID, useExpanderCell } from './useExpanderCell';
 export { SELECTION_CELL_ID, useSelectionCell } from './useSelectionCell';
+export { useSubRowFiltering } from './useSubRowFiltering';
+export { useSubRowSelection } from './useSubRowSelection';