+++ stories/core/Table.stories.tsx
@@ -39,6 +39,7 @@ export default {
     ],
     emptyTableContent: 'No data.',
     density: 'default',
+    emptyFilteredTableContent: 'No results found. Clear or try another filter.',
   },
   argTypes: {
     columns: { control: { disable: true } },
@@ -486,7 +487,6 @@ Filters.args = {
       endDate: '2021-06-02T21:00:00.000Z',
     },
   ],
-  emptyFilteredTableContent: 'No results found. Clear or try another filter.',
 };
 Filters.parameters = {
   creevey: {
@@ -601,6 +601,132 @@ Expandable.parameters = {
         await expanderButtons[2].click();
 
         const expanded = await this.takeScreenshot();
+        await this.expect({ closed, expanded }).to.matchImages();
+      },
+    },
+  } as CreeveyStoryParams,
+};
+
+export const ExpandableSubrows: Story<TableProps> = (args) => {
+  const { data, ...rest } = args;
+
+  const onExpand = useCallback(
+    (rows, state) =>
+      action(
+        `Expanded rows: ${JSON.stringify(rows)}. Table state: ${JSON.stringify(
+          state,
+        )}`,
+      )(),
+    [],
+  );
+
+  const tableColumns = useMemo(
+    () => [
+      {
+        Header: 'Table',
+        columns: [
+          {
+            id: 'name',
+            Header: 'Name',
+            accessor: 'name',
+            Filter: tableFilters.TextFilter(),
+          },
+          {
+            id: 'description',
+            Header: 'Description',
+            accessor: 'description',
+          },
+        ],
+      },
+    ],
+    [],
+  );
+
+  return (
+    <Table
+      emptyTableContent='No data.'
+      isSelectable
+      isSortable
+      {...rest}
+      data={data}
+      columns={tableColumns}
+      onExpand={onExpand}
+    />
+  );
+};
+
+ExpandableSubrows.args = {
+  data: [
+    {
+      name: 'Row 1',
+      description: 'Description 1',
+      subRows: [
+        { name: 'Row 1.1', description: 'Description 1.1', subRows: [] },
+        {
+          name: 'Row 1.2',
+          description: 'Description 1.2',
+          subRows: [
+            {
+              name: 'Row 1.2.1',
+              description: 'Description 1.2.1',
+              subRows: [],
+            },
+            {
+              name: 'Row 1.2.2',
+              description: 'Description 1.2.2',
+              subRows: [],
+            },
+            {
+              name: 'Row 1.2.3',
+              description: 'Description 1.2.3',
+              subRows: [],
+            },
+            {
+              name: 'Row 1.2.4',
+              description: 'Description 1.2.4',
+              subRows: [],
+            },
+          ],
+        },
+        { name: 'Row 1.3', description: 'Description 1.3', subRows: [] },
+        { name: 'Row 1.4', description: 'Description 1.4', subRows: [] },
+      ],
+    },
+    {
+      name: 'Row 2',
+      description: 'Description 2',
+      subRows: [
+        { name: 'Row 2.1', description: 'Description 2.1', subRows: [] },
+        { name: 'Row 2.2', description: 'Description 2.2', subRows: [] },
+        { name: 'Row 2.3', description: 'Description 2.3', subRows: [] },
+      ],
+    },
+    { name: 'Row 3', description: 'Description 3', subRows: [] },
+  ],
+};
+ExpandableSubrows.argTypes = {
+  data: { control: { disable: true } },
+};
+
+ExpandableSubrows.parameters = {
+  creevey: {
+    tests: {
+      async expand() {
+        const closed = await this.takeScreenshot();
+
+        let expanders = await this.browser.findElements({
+          css: '.iui-row-expander',
+        });
+        // Expand Row 1
+        await expanders[0].click();
+        expanders = await this.browser.findElements({
+          css: '.iui-row-expander',
+        });
+        // Expand Row 1.2
+        await expanders[1].click();
+
+        const expanded = await this.takeScreenshot();
+
         await this.expect({ closed, expanded }).to.matchImages();
       },
     },